+++ src/main/scala/io/iohk/ethereum/keystore/Wallet.scala
@@ -8,6 +8,6 @@ import org.bouncycastle.crypto.AsymmetricCipherKeyPair
 case class Wallet(address: Address, prvKey: ByteString) {
   lazy val keyPair: AsymmetricCipherKeyPair = keyPairFromPrvKey(prvKey.toArray)
 
-  def signTx(tx: Transaction, chainId: Option[BigInt]): SignedTransactionWithSender =
+  def signTx(tx: Transaction, chainId: Option[Byte]): SignedTransactionWithSender =
     SignedTransaction.sign(tx, keyPair, chainId)
 }
