+++ src/main/scala/io/iohk/ethereum/network/rlpx/AuthInitiateEcdsaCodec.scala
@@ -26,6 +26,6 @@ trait AuthInitiateEcdsaCodec {
     val r = input.take(RLength)
     val s = input.slice(SIndex, SIndex + SLength)
     val v = input(VIndex) + 27
-    ECDSASignature(BigInt(1, r), BigInt(1, s), v)
+    ECDSASignature(BigInt(1, r), BigInt(1, s), v.toByte)
   }
 }
